pyproject.toml
[project]
name = "floorplan-llm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "opencv-python-headless>=4.13.0.92",
    "openpyxl>=3.1.5",
    "orjson>=3.11.7",
    "pillow>=12.1.1",
    "torch>=2.6.0",
    "transformers>=4.51.0",
    "tqdm>=4.67.3",
    "datasets>=4.6.1",
    # Pre-Stage 훈련 의존성
    "peft>=0.15.0",
    "bitsandbytes>=0.45.0",
    "accelerate>=1.6.0",
    "wandb>=0.19.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
